pyproject.toml
[project]
name = "personal-assistant-with-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.6",
    "langchain-text-splitters>=0.3.8",
    "pymupdf4llm>=0.0.26",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
